♻️ [Refactor]: 커리큘럼 FSD 구조 리팩토링
