+++ src/widget/study-room/curriculum/ui/CurriculumForm.tsx
@@ -1,30 +1,33 @@
-import { memo, useCallback, useRef } from "react";
-import { useFormContext } from "react-hook-form";
+import { memo } from "react";
+import { FieldError } from "react-hook-form";
 
-import { DndProvider, useDrag, useDrop } from "react-dnd";
+import { DndProvider } from "react-dnd";
 import { HTML5Backend } from "react-dnd-html5-backend";
 
-import { StudyCurriculumCardInterface } from "@/features/study-room/types/study-room.type";
 import Card from "@/shared/components/atoms/Card";
 import Input from "@/shared/components/atoms/Input";
 import Tag from "@/shared/components/atoms/Tag";
 import Textarea from "@/shared/components/atoms/Textarea";
 import Typography from "@/shared/components/atoms/Typography";
-
-const ItemType = "CURRICULUM_ITEM";
+import { StudyCurriculumCardInterface } from "@/entities/study/curriculum/model/curriculum.types";
+import useDragCurriculum from "@/features/curriculum/dnd/model/useDragCurriculum";
+import useDeleteCurriculum from "@/features/curriculum/delete-curriculum/model/useDeleteCurriculum";
+import Button from "@/shared/components/atoms/Button";
+import useAddCurriculum from "@/features/curriculum/add-curriculum/model/useAddCurriculum";
+import { DeleteButton } from "@/features/curriculum/delete-curriculum/ui/DeleteButton";
+import useUpdateCurriculum from "@/features/curriculum/update-curriculum/model/useUpdateCurriculum";
+import { AddButton } from "@/features/curriculum/add-curriculum/ui/AddButton";
 
 interface CurriculumProps {
-  isModify: boolean;
+  studyId: number;
 }
 
 interface CurriculumItemProps {
   index: number;
   curriculum: StudyCurriculumCardInterface;
+  studyId: number;
+  hasAnyFieldError?: (index: number) => FieldError | undefined;
 }
-
-type FormValues = {
-  curriculumList: StudyCurriculumCardInterface[];
-};
 
 const CurriculumView = ({ index, curriculum }: CurriculumItemProps) => (
   <div className="flex gap-[20px]" key={index}>
@@ -58,38 +61,12 @@ const CurriculumView = ({ index, curriculum }: CurriculumItemProps) => (
   </div>
 );
 
-const CurriculumEditView = ({ index, curriculum }: CurriculumItemProps) => {
-  // react-hook-form
-  const {
-    register,
-    getValues,
-    setValue,
-    formState: { errors },
-  } = useFormContext<FormValues>();
-  const curriculumList = getValues("curriculumList");
-
-  // 커리큘럼 삭제
-  const deleteCurriculum = useCallback(
-    (id: number) => {
-      const filteredList = curriculumList.filter((item) => item.id !== id);
-
-      // sectionId를 1부터 순차적으로 재정렬
-      const updatedList = filteredList.map((item, index) => ({
-        ...item,
-        sectionId: index + 1,
-      }));
-
-      setValue("curriculumList", updatedList);
-    },
-
-    [curriculumList, setValue]
-  );
-
-  // step, title, content 중 하나라도 에러가 있으면 true 에러가 없으면 false
-  const hasAnyFieldError = (index: number) => {
-    const error = errors?.curriculumList?.[index];
-    return error?.sectionId || error?.title || error?.content;
-  };
+const CurriculumEditView = ({
+  index,
+  curriculum,
+  hasAnyFieldError,
+}: CurriculumItemProps) => {
+  const { register } = useDeleteCurriculum();
 
   return (
     <div className="flex gap-[5px]">
@@ -120,11 +97,8 @@ const CurriculumEditView = ({ index, curriculum }: CurriculumItemProps) => {
               defaultValue={curriculum.title}
               className="h-[30px] w-[99%]"
             />
-            {/* 삭제 아이콘 */}
-            <i
-              className="bi bi-trash3 cursor-pointer text-mos-coral-500"
-              onClick={() => deleteCurriculum(curriculum.id!)}
-            />
+            {/* 삭제 버튼 */}
+            <DeleteButton sectionId={curriculum.sectionId!} />
           </Card.Header>
 
           <Card.Content className="flex flex-col gap-1">
@@ -137,7 +111,7 @@ const CurriculumEditView = ({ index, curriculum }: CurriculumItemProps) => {
                 defaultValue={curriculum.content}
               />
               <Typography.Error>
-                {hasAnyFieldError(index) &&
+                {hasAnyFieldError!(index) &&
                   "해당 커리큘럼 항목의 모든 정보를 입력해주세요."}
               </Typography.Error>
             </div>
@@ -152,95 +126,54 @@ const DraggableCurriculumEditView = ({
   index,
   curriculum,
   moveItem,
+  studyId,
+  hasAnyFieldError,
 }: CurriculumItemProps & {
   moveItem: (dragIndex: number, hoverIndex: number) => void;
 }) => {
-  // DnD를 연결할 대상
-  const ref = useRef<HTMLDivElement>(null);
-
-  // 해당 항목이 드롭 대상이 되도록 설정
-  const [, drop] = useDrop({
-    accept: ItemType,
-    // 이 항목이 드래그된 요소가 위에 올라왔을 때 반응할 수 있도록 설정
-    hover(item: { index: number }, monitor) {
-      if (!ref.current) return;
-      const dragIndex = item.index;
-      const hoverIndex = index;
-
-      if (dragIndex === hoverIndex) return; // 위치가 그대로면 무시
-
-      //  마우스 위치 계산
-      const hoverBoundingRect = ref.current?.getBoundingClientRect();
-      const hoverMiddleY =
-        (hoverBoundingRect.bottom - hoverBoundingRect.top) / 2;
-      const clientOffset = monitor.getClientOffset();
-      const hoverClientY = clientOffset!.y - hoverBoundingRect.top;
-
-      if (dragIndex < hoverIndex && hoverClientY < hoverMiddleY) return;
-      if (dragIndex > hoverIndex && hoverClientY > hoverMiddleY) return;
-
-      // 위 조건을 통과하면 실제로 moveItem 호출해서 순서 바꿈
-      moveItem(dragIndex, hoverIndex);
-      item.index = hoverIndex;
-    },
-  });
-
-  // 해당 항목이 드래그 가능한 항목이 되도록 설정
-  const [{ isDragging }, drag] = useDrag({
-    type: ItemType,
-    item: { index },
-    collect: (monitor) => ({
-      isDragging: monitor.isDragging(), // isDragging: 드래그 항목의 투명도 조절 기능
-    }),
-  });
-
-  drag(drop(ref));
+  const { ref, isDragging } = useDragCurriculum(index, moveItem);
 
   return (
     <div ref={ref} style={{ opacity: isDragging ? 0.5 : 1 }}>
-      <CurriculumEditView index={index} curriculum={curriculum} />
+      <CurriculumEditView
+        index={index}
+        curriculum={curriculum}
+        studyId={studyId}
+        hasAnyFieldError={hasAnyFieldError}
+      />
     </div>
   );
 };
 
-const Curriculum = ({ isModify }: CurriculumProps) => {
-  const { watch, setValue } = useFormContext<{
-    curriculumList: StudyCurriculumCardInterface[];
-  }>();
-  const curriculumList = watch("curriculumList");
+const Curriculum = ({ studyId }: CurriculumProps) => {
+  const {
+    setIsCreateState,
+    curriculumList,
+    isValidCurriculum,
+    hasAnyFieldError,
+  } = useAddCurriculum();
 
-  // dnd 아이템 이동
-  // dragIndex: 드래그한 아이템의 인덱스 (클릭한 항목)
-  // hoverIndex: 드래그한 아이템을 놓을 목표 인덱스
-  const moveItem = (dragIndex: number, hoverIndex: number) => {
-    const updatedList = [...curriculumList];
-    const [draggedItem] = updatedList.splice(dragIndex, 1); // draggedItem: 기존 배열에서 이동시킬 아이템을 뽑아옴
-    updatedList.splice(hoverIndex, 0, draggedItem); // 기존 배열의 hoverIndex 위치에 draggedItem을 삽입
-    // sectionId를 index 순서에 맞게 재정렬
-    const reorderedList = updatedList.map((item, index) => ({
-      ...item,
-      sectionId: index + 1,
-    }));
-
-    setValue("curriculumList", reorderedList);
-  };
+  const { isModifyState, setIsModifyState, moveItem } = useUpdateCurriculum();
 
   const renderItem = (
     curriculum: StudyCurriculumCardInterface,
     index: number
   ) =>
-    isModify ? (
+    isModifyState ? (
       <DraggableCurriculumEditView
         key={curriculum.sectionId}
         index={index}
         curriculum={curriculum}
         moveItem={moveItem}
+        studyId={studyId}
+        hasAnyFieldError={hasAnyFieldError}
       />
     ) : (
       <CurriculumView
         key={curriculum.sectionId}
         index={index}
         curriculum={curriculum}
+        studyId={studyId}
       />
     );
 
@@ -253,6 +186,29 @@ const Curriculum = ({ isModify }: CurriculumProps) => {
           </Typography.P3>
         </div>
       )}
+      <div className="mb-4 flex items-center justify-between">
+        <Typography.SubTitle1>커리큘럼</Typography.SubTitle1>
+        <div className="flex gap-2">
+          {isModifyState && <AddButton />}
+
+          <Button.Solid
+            color="Main"
+            className="h-[30px] text-[14px]"
+            onClick={() => {
+              setIsCreateState(true);
+              setIsModifyState((prev) => !prev);
+            }}
+            type={isModifyState ? "button" : "submit"}
+            active={isValidCurriculum}
+          >
+            {curriculumList.length === 0
+              ? "등록"
+              : isModifyState
+              ? "확인"
+              : "수정"}
+          </Button.Solid>
+        </div>
+      </div>
       {curriculumList.map((curriculum, index) => renderItem(curriculum, index))}
     </DndProvider>
   );
